DeepSceneAI/requirements.txt
# ===============================
# Core Machine Learning & AI
# ===============================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
diffusers>=0.26.0
accelerate>=0.24.0

# ===============================
# AI Image Generation (FREE Local)
# ===============================
# Required for Stable Diffusion
safetensors>=0.4.0
xformers>=0.0.22              # Optional: for faster GPU generation
omegaconf>=2.3.0              # Configuration management

# ===============================
# Data Processing & Analysis
# ===============================
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ===============================
# Natural Language Processing
# ===============================
spacy>=3.7.0
# After installing, run: python -m spacy download en_core_web_sm
nltk>=3.8.0

# ===============================
# Web Framework & API
# ===============================
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# ===============================
# Image Processing
# ===============================
Pillow>=10.0.0
opencv-python>=4.8.0          # Optional: for advanced image processing
imageio>=2.31.0

# ===============================
# HTTP Requests & Web
# ===============================
requests>=2.31.0
aiohttp>=3.9.0                # Optional: for async requests
httpx>=0.25.0

# ===============================
# Development & Utilities
# ===============================
jupyterlab>=4.0.0
notebook>=7.0.0
ipykernel>=6.27.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.66.0
python-dotenv>=1.0.0          # Environment variables

# ===============================
# Testing & Quality
# ===============================
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0                  # Code formatting
flake8>=6.0.0                  # Linting

# ===============================
# File & Data Management
# ===============================
pyyaml>=6.0.0
python-json-logger>=2.0.0
pathlib2>=2.3.0; python_version < "3.4"

# ===============================
# Optional: Cloud & APIs
# ===============================
# openai>=1.3.0                 # Optional: for GPT-4/DALL-E (paid)
# boto3>=1.28.0                 # Optional: for AWS services
# google-cloud-aiplatform>=1.38.0 # Optional: for Google AI

# ===============================
# System & Performance
# ===============================
psutil>=5.9.0                  # System monitoring
GPUtil>=1.4.0                  # GPU monitoring